creating init
