+++ src/student_success_tool/modeling/h2o/__init__.py
@@ -1,0 +1,4 @@
+from . import (
+    training,
+    evaluation,
+)
