+++ src/student_success_tool/modeling/h2o/training.py
@@ -119,7 +119,7 @@ def log_h2o_experiment(
     if client is None:
         client = MlflowClient()
 
-    experiment_id = training_utils.set_or_create_experiment(
+    experiment_id = set_or_create_experiment(
         workspace_path,
         institution_id,
         target_name,
@@ -176,3 +176,35 @@ def log_h2o_experiment(
     LOGGER.info(f"Logged {len(results_df)} model runs to MLflow.")
 
     return experiment_id, results_df
+
+
+def set_or_create_experiment(workspace_path, institution_id, target_name, checkpoint_name, use_timestamp=True, client=None):
+    if client is None:
+        client = MlflowClient()
+
+    timestamp = datetime.datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%S") if use_timestamp else ""
+
+    name_parts = [
+        cfg.institution_id,
+        cfg.preprocessing.target.name,
+        cfg.preprocessing.checkpoint.name,
+        prefix,
+        timestamp
+    ]
+    experiment_name = "/".join([
+        workspace_path.rstrip("/"),
+        "h2o_automl",
+        "_".join([part for part in name_parts if part]),
+    ])
+
+    try:
+        experiment = client.get_experiment_by_name(experiment_name)
+        if experiment is None:
+            experiment_id = client.create_experiment(experiment_name)
+        else:
+            experiment_id = experiment.experiment_id
+
+        mlflow.set_experiment(experiment_name)
+        return experiment_id
+    except Exception as e:
+        raise RuntimeError(f"Failed to create or set MLflow experiment: {e}")
